fix: move updating categories for market page into a formatting function.
